7.2* wrong overflowcheck on ValRawInt64 and ValUnicodeInt64
git-svn-id: http://svn.code.sf.net/p/zeoslib/code-0/branches/testing-7.3@5349 d681b350-854c-439c-a81d-c0ce20710203

+++ src/core/ZFastCode.pas
@@ -4328,7 +4328,7 @@ begin
       if ((Flags and 2) = 0) or (UInt64(Result) <> UInt64(Int64($8000000000000000))) then
     {$ELSE}
     if UInt64(Result) >= $8000000000000000 then {Possible Overflow}
-      if ((Flags and 2) = 0) or (Result <> $8000000000000000) then
+      if ((Flags and 2) = 0) or (UInt64(Result) <> $8000000000000000) then
     {$ENDIF}
         begin {Overflow}
           if ((Flags and 2) <> 0) then {Neg=True}
@@ -4871,7 +4871,7 @@ begin
       Inc(P);
     end;
     if UInt64(Result) >= {$IFDEF FPC}HighInt64{$ELSE}$8000000000000000{$ENDIF} then {Possible Overflow}
-      if ((Flags and 2) = 0) or (Result <> {$IFDEF FPC}HighInt64{$ELSE}$8000000000000000{$ENDIF}) then begin {Overflow}
+      if ((Flags and 2) = 0) or (UInt64(Result) <> {$IFDEF FPC}HighInt64{$ELSE}$8000000000000000{$ENDIF}) then begin {Overflow}
         if ((Flags and 2) <> 0) then {Neg=True}
           Result := -Result;
         Code := P-S;
